circle-6/ft_transcendence/backend/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
psycopg2-binary==2.9.1
django-cors-headers==3.8.0
channels==4.0.0
daphne==4.0.0
Pillow==9.0

